% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fwindow}
\alias{fwindow}
\title{Window clone}
\usage{
fwindow(x, start = NA_real_, end = NA_real_, ...)
}
\arguments{
\item{x}{a time-series (or other object if not replacing values)}

\item{start}{the start time of the period of interest.}

\item{end}{the end time of the period of interest.}

\item{...}{ignored}
}
\value{
a time-serie
}
\description{
Window clone
}
